% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{cols}
\alias{cols}
\title{Extracts the vector of colours from a pdata object}
\usage{
cols(pdata)
}
\arguments{
\item{pdata}{a pdata object consisting of a list including a data frame called 'pdata',
containing phenotypic information (on columns) for samples (on rows). One of those is a
column called 'cols_vec' containing colours that are to be used in plots, corresponding
to the samples.}
}
\value{
A character vector containing colours corresponding to samples in the pdata object.
}
\description{
This function extracts the column 'cols_vec' from the 'pdata slot of
a given pdata object.
}
