% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdata.R
\name{combine.pdata}
\alias{combine.pdata}
\alias{combinePdata}
\title{Combines pdata objects corresponding to multiple screens}
\usage{
combine.pdata(list_pdata)
}
\arguments{
\item{list_pdata}{a list of pdata objects, read in using get.pdata}
}
\value{
A list containing the stacked pdata objects, a single colour variable coding for the cell line variable, as well as
a character vector with the variable names containing replicate names, cell line ids and treatment/condition.
}
\description{
\code{combine.pdata} combines pdata from multiple screens, already read in
using get.pdata into a single pdata object. Variables are combined to allow for
using them for plots later on. Variable names are standardized to allow for
combining all observations smoothly. Other variables that exist in one pdata and not in
the other, receive NA for the extra observations.
}
\examples{
# See vignette
}
\seealso{
\code{\link{read.screen.data}} to read in screen data, \code{\link{combine.screens}}
to combine data from multiple screens,  and \code{\link{get.leth.scores}} to compute scores
that make observations in different screens comparable.
}
