% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_screen_data.R
\name{combine.screens}
\alias{combine.screens}
\title{Combines different screen data objects into a single one.}
\usage{
combine.screens(list_rscreen)
}
\arguments{
\item{list_rscreen}{a list of rscreen objects, such as the ones produced by reading
data with read.screen.data.}
}
\value{
An object of class rscreen.object.
}
\description{
\code{combine.screens} combines data from multiple screens, already read in
using read.screen.data into a single rscreen.object. The different screens
involve the same set of library features, but may use different plate sets and/or
different (numbers of) controls. As such, different screens share annotation
relating to the library features, such as (target) gene id, but are likely to
have different plate and/or well annotation. The total number of features
measured per screen must be the same for all experiments.
}
\examples{
# See vignette
}
\seealso{
\code{\link{read.screen.data}} to read data for individual screens,
\code{\link{combine.pdata}} for combining phenotypic data tables corresponding
to different screens that are to be normalized together, and
\code{\link{get.leth.scores}} to compute lethality scores for data corresponding
to multiple screens
}
