% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lethScores.R
\name{comp.scores.diff.ann}
\alias{comp.scores.diff.ann}
\title{Computes lethality scores for replicates with different plate/well annotation}
\usage{
comp.scores.diff.ann(my_data, my_ann, is_robust = TRUE)
}
\arguments{
\item{my_data}{data matrix containing only data columns, so all columns are of type numeric.
It can be a matrix with a single column, but we must be able to apply nrow, ncol and apply on it.}

\item{my_ann}{list with data frames, each with as many columns as those in \code{my_data} and
with the same column names, containing well annotation (wtype_ann) and plate (plate_ann).}

\item{is_robust}{logical; \code{TRUE} by default, when the median of the controls is used to
compute the scores. If \code{FALSE}, the mean of the controls is used instead.}
}
\value{
a matrix with the same dimensions as \code{my_data} containing all lethality scores.
}
\description{
\code{comp.scores.diff.ann} is an internal function called by get.leth.scores.
It computes lethality scores per replicate and plate in a data matrix, for which plate and well
annotation may vary across replicates.
}
\seealso{
\code{\link{read.screen.data}} on reading screen data, \code{\link{combine.screens}}
to combine data from multiple screens into a single object, and \code{\link{get.rscreenorm}} to
normalize the screen data using the invariant-set quantile normalization.
}
