% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lethScores.R
\name{comp.scores}
\alias{comp.scores}
\title{Computes lethality scores for replicates with the same plate/well annotation}
\usage{
comp.scores(my_data, wtype_ann, is_robust = TRUE)
}
\arguments{
\item{my_data}{data matrix containing only data columns, so all columns are of type numeric.
It can be a matrix with a single column, but we must be able to apply nrow, ncol and apply on it.}

\item{wtype_ann}{object containing the well type annotation. It can be a vector, character or factor.
In case different columns of \code{my_data} have different well annotation, this should be a matrix
with the same dimensions as \code{my_data}.}

\item{is_robust}{logical; \code{TRUE} by default, when the median of the controls is used to
compute the scores. If \code{FALSE}, the mean of the controls is used instead.}
}
\value{
a matrix with the same dimensions as \code{my_data} containing all lethality scores.
}
\description{
\code{comp.scores} is an internal function called by get.leth.scores.
It computes lethality scores per replicate, for a data matrix where replicates
have the same plate/well annotation.
}
\seealso{
\code{\link{read.screen.data}} on reading screen data, \code{\link{combine.screens}}
to combine data from multiple screens into a single object, and \code{\link{get.rscreenorm}} to
normalize the screen data using the invariant-set quantile normalization.
}
