% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qnorm.R
\name{get.all.qnorm}
\alias{get.all.qnorm}
\alias{getallqnorm}
\title{Computes normalized data for all observed scores}
\usage{
get.all.qnorm(data_rscreen, data_perc, data_perc_qn)
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', most likely corresponding to lethality scores
produced by \code{\link{get.leth.scores}}. It is also possible to normalize re-scaled screen
viability data, such as yielded by \code{\link{read.screen.data}}, or by \code{\link{combine.screens}},
although these tend to be less comparable between cell lines and replicates than the scores.}

\item{data_perc}{the percentiles of the per-replicate scores for the replicate-specific
invariannt sets, as yielded by  \code{\link{get.perc}}.}

\item{data_perc_qn}{the quantile-normalized percentiles of the per-replicate data, as yielded by
\code{\link{get.perc.qnorm}}.}
}
\value{
a matrix with the same dimensions as the 'data_only' slot of \code{data_rscreen}, containing the
quantile-normalized scores.
}
\description{
This is an internal function that yields normalized lethality scores
by fitting a regression between the
original and the normalized percentiles for core set scores. It is
called by both \code{\link{get.rscreenorm}} and \code{\link{get.qnorm}}, and is to be used
internally.
}
\seealso{
\code{\link{get.rscreenorm}} which yields core sets and quantile-normalizes data at
once, code{\link{get.qnorm}} which quantile-normalizes data for a given core set,
\code{\link{get.perc}} to compute percentiles for the core set per replicate, \code{\link{get.inv.set}} for
obtaining core sets for all replicates in the dataset and
\code{\link{get.perc.qnorm}} to compute the quantile-normalized percentiles.
}
\concept{
scores screen siRNA CRISPR CRISPR-Cas9 crispr-cas9 CRISPR/Cas9 crispr/cas9 quantile normalization
}
