% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_set.R
\name{get.inv.set.prop}
\alias{get.inv.set.prop}
\alias{getInvSetProp}
\title{Computes proportion of scores included in core sets}
\usage{
get.inv.set.prop(data_rscreen, th_obj, plot = FALSE, cols_list = NULL)
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', either yielded by \code{\link{read.screen.data}}, or
by \code{\link{combine.screens}}. Typically the 'data_only' slot of this object
contains lethality scores computed by \code{\link{get.leth.scores}}, which already
reflect a value relative to negative and positive controls, per replicate.}

\item{th_obj}{object containing thresholds per replicate, as returned by
\code{\link{get.th.inv.set}}.}

\item{plot}{logical, indicating if a barplot is to be made of the result. Defaults
to \code{FALSE}, when no plot is produced.}

\item{cols_list}{vector of colours to be used for the barplot, so if given it must
be as long as the number of columns in data_only slot of data_rscreen. If \code{\link{get.pdata}}
was used to read the phenotypic data table, its slot cols_list can be used as input
for this argument. By default, a vector of colours created by \code{\link{rainbow}} will
be used, a different one for each column in data_only.}
}
\value{
A vector as long as the number of columns in the data_only slot of data_rscreen,
with each entry corresponding to the proportion of scores included in an core set.
If plot = \code{TRUE}, makes also a barplot of the computed proportions.
}
\description{
This function uses data_rscreen and a threshold object and returns a
vector of proportions of scores included in core sets, per replicate. It can also
create a barplot of the computed proportions.
It is typically used when a \code{my_gamma} is chosen to select core sets,
yielding a check on the built sets.
}
\examples{
# See vignette
}
\seealso{
\code{\link{get.inv.set}} the wrapper function that yields core sets,
\code{\link{get.th.inv.set}} that produces the thresholds per replicate, \code{\link{get.rscreenorm}}
to normalize data from multiple screens using core sets and \code{\link{get.leth.scores}}
to compute scores that make observations in different screens comparable.
}
\concept{
core set normalization quantile threshold
}
