% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_set.R
\name{get.inv.set}
\alias{get.inv.set}
\alias{getInvSet}
\title{Wrapper where the core set is created}
\usage{
get.inv.set(data_rscreen, my_gamma = NULL, var_type = c("mad", "IQR", "sd"),
  prop = 0.95, set_shape = c("left", "centre"))
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', either yielded by \code{\link{read.screen.data}}, or
by \code{\link{combine.screens}}. Typically the 'data_only' slot of this object
contains lethality scores computed by \code{\link{get.leth.scores}}, which already
reflect a value relative to negative and positive controls, per replicate.}

\item{my_gamma}{scalar or numeric vector, with value(s) between 0 and a finite
number. Its default value is \code{NULL}. If given, it will be used to construct the
core sets by using the relative distance between negative and positive
controls' distributions. If a numeric vector is given, it must contain as many
entries as columns in the data_only slot of data_rscreen. A value of 1 means that
the core set includes all lethality scores up to the ratio between the MAD
of negative controls, and the sum of the MADs of negative and positive controls,
assuming the robust variability measure MAD is used.}

\item{var_type}{string indicating the statistic used as variability measure.
Possible values are 'mad' when the median absolute deviation (MAD) is used
(the default), 'IQR' when the inter-quartile range is used, and "sd" when the
standard deviation is used.}

\item{prop}{scalar between 0 and 1, corresponding to the desired proportion of
lethality scores to be included in the core set, per replicate. The default
value is 0.95. If \code{my_gamma} is numeric, any value for this argument is ignored.}

\item{set_shape}{string indicating the shape of the core set. It accepts
"left" (the default), when all scores to the left of the set threshold are
included, per replicate, and "centre", when scores included in the core
set only exclude the largest and smallest ones, at equal frequencies.}
}
\value{
A logical matrix with the same dimensions as the data_only slot of data_rscreen.
}
\description{
This function yields one core set of lethality scores per replicate, for
a given rscreen.object. The core sets are used as basis for quantile-normalizing
data for independent screens and their replicates.
}
\examples{
# See vignette
}
\seealso{
\code{\link{get.rscreenorm}} to normalize data from multiple screens using core
sets and \code{\link{get.leth.scores}} to compute scores that make observations in
different screens comparable.
}
\concept{
core set normalization quantile
}
