% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lethScores.R
\name{get.leth.scores}
\alias{get.leth.scores}
\alias{getLethScores}
\title{Computes lethality scores per replicate in an rscreen.data object}
\usage{
get.leth.scores(data_rscreen, robust = TRUE, rescale = FALSE,
  scale.fun = c("asinh", "log2"))
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', either yielded by \code{\link{read.screen.data}}, or
by \code{\link{combine.screens}}. Typically the 'data_only' slot of this object
contains lethality scores computed by \code{\link{get.leth.scores}}, which already
reflect a value relative to negative and positive controls, per replicate.}

\item{robust}{logical; \code{TRUE} by default, when the median of the controls is used to
compute the scores. If \code{FALSE}, the mean of the controls is used instead.}

\item{rescale}{logical, indicating whether or not the data is to be rescaled. Defaults to \code{FALSE}.
If \code{TRUE}, the screen data will be rescaled per replicate using the function indicated in 'scale.fun'.}

\item{scale.fun}{the name of the function to be used to rescale the data. Currently the
only choices are 'asinh' for the hyperbolic-arc sine transformation, and 'log2' for the
log2 transformation. Note that the hyperbolic-arc
sine is equivalent to the logarithm for intermediate and high values, and to a linear
transformation for low values. This transformation allows for zeros, which is a useful property when
transforming count data. Ignored if 'rescale' is \code{FALSE}.}
}
\value{
An object of class rscreen.object, with its data_only slot containing lethality scores.
}
\description{
\code{get.leth.scores} computes lethality scores per replicate in the \code{data_rscreen}
object. The \code{data_rscreen} object  may either involve data either for a single screen, or
for multiple screens. Note that lethality scores are computed per plate, if a plate variable
is available in the 'rscreen.object' given as input. This corrects automatically for any plate
effect that affects all wells in a plate to the same extent.
}
\examples{
# See vignette
}
\seealso{
\code{\link{read.screen.data}} on reading screen data, \code{\link{combine.screens}}
to combine data from multiple screens into a single object, and \code{\link{get.rscreenorm}} to
normalize the screen data using the invariant-set quantile normalization.
}
\concept{
scores screen siRNA CRISPR CRISPR-Cas9 crispr-cas9 CRISPR/Cas9 crispr/cas9
}
