% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lethScores.R
\name{get.list.res}
\alias{get.list.res}
\title{Selects lethality scores subsets per feature type}
\usage{
get.list.res(data_only, wtype_mat)
}
\arguments{
\item{data_only}{data matrix containing only data columns, so all columns are of type numeric.}

\item{wtype_mat}{data.frame or matrix with the same dimensions as data_only, where column \emph{k}
contains well annotation for column \emph{k} in data_only.}
}
\value{
a list consisting of mat_let, a matrix with lethality scores corresponding to library
features; let_neg, a list with as many elements as columns in mat_sample, with negative controls'
lethality scores; let_pos, a list with as many elements as columns in mat_sample, with positive
controls' lethality scores
}
\description{
\code{get.list.res} is an internal function called by functions involved in constructing
invariant sets. It selects subsets of lethality scores per replicate and per feature
type, namely library (or sample) feature, positive and negative controls.
}
\seealso{
\code{\link{read.screen.data}} on reading screen data, \code{\link{combine.screens}}
to combine data from multiple screens into a single object, and \code{\link{get.rscreenorm}} to
normalize the screen data using the invariant-set quantile normalization.
}
