% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qnorm.R
\name{get.perc}
\alias{get.perc}
\alias{getperc}
\title{Computes percentiles for observations in the core set}
\usage{
get.perc(data_rscreen, inv_set, n_perc = 1000)
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', most likely corresponding to lethality scores
produced by \code{\link{get.leth.scores}}. It is also possible to normalize re-scaled screen
viability data, such as yielded by \code{\link{read.screen.data}}, or by \code{\link{combine.screens}},
although these tend to be less comparable between cell lines and replicates than the scores.}

\item{inv_set}{a logical matrix with the same dimensions as the 'data_only' slot in \code{data_rscreen},
indicating the observations that belong to the core set. This is typically the output of
\code{\link{sel.scores.inv.set}}.}

\item{n_perc}{number of percentiles to be used to yield a representation of each
core set. The default value is 1000, which is appropriate for pooled, whole
genome screens. For small, hit-picking screens a smaller number may be more appropriate.
The user may check the core set sizes to decide if in doubt. Note that the value
of \code{n_perc}  should not be larger than that of any core set.}
}
\value{
a matrix with as many rows as n_perc and as many columns as in the data_only slot
of data_rscreen, containing the percentiles for the core sets, per replicate.
}
\description{
This is an internal function that
computes percentiles for the data in the core set. This function is
called by both \code{\link{get.rscreenorm}} and \code{\link{get.qnorm}}. It is to be used
internally, so has no functionality other than what is described.
}
\seealso{
\code{\link{get.rscreenorm}} which yields core sets and quantile-normalizes data at
once, code{\link{get.qnorm}} which quantile-normalizes data for a given core set, \code{\link{get.inv.set}} for
obtaining core sets for all replicates in the dataset and \code{\link{get.leth.scores}} to compute lethality scores.
}
\concept{
scores screen siRNA CRISPR CRISPR-Cas9 crispr-cas9 CRISPR/Cas9 crispr/cas9 quantile normalization
}
