% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qnorm.R
\name{get.qnorm}
\alias{get.qnorm}
\alias{getqnorm}
\title{Quantile-normalizes screen data using given core sets}
\usage{
get.qnorm(data_rscreen, inv_set, n_perc = 1000)
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', most likely corresponding to lethality scores
produced by \code{\link{get.leth.scores}}. It is also possible to normalize re-scaled screen
viability data, such as yielded by \code{\link{read.screen.data}}, or by \code{\link{combine.screens}},
although these tend to be less comparable between cell lines and replicates than the scores.}

\item{inv_set}{a logical matrix with the same dimensions as the 'data_only' slot in \code{data_rscreen},
indicating the observations that belong to the core set. This is typically the output of
\code{\link{sel.scores.inv.set}}.}

\item{n_perc}{number of percentiles to be used to yield a representation of each
core set. The default value is 1000, which is appropriate for pooled, whole
genome screens. For small, hit-picking screens a smaller number may be more appropriate.
The user may check the core set sizes to decide if in doubt. Note that the value
of \code{n_perc}  should not be larger than that of any core set.}
}
\value{
An object of class rscreen.object, with its "data_only" slot containing quantile-normalized
lethality scores.
}
\description{
This is a wrapper for multiple functions, that will for given core sets yield
quantile-normalize scores for all replicates.
}
\examples{
# See vignette
}
\seealso{
\code{\link{get.rscreenorm}} which yields core sets and quantile-normalizes data at
once, \code{\link{get.inv.set}} for obtaining core sets for all replicates in the dataset
and \code{\link{get.leth.scores}} to compute lethality scores.
}
\concept{
scores screen siRNA CRISPR CRISPR-Cas9 crispr-cas9 CRISPR/Cas9 crispr/cas9 quantile normalization
}
