% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qnorm.R
\name{get.rscreenorm}
\alias{get.rscreenorm}
\alias{getrscreenorm}
\title{Quantile-normalizes screen data using core sets}
\usage{
get.rscreenorm(data_rscreen, my_gamma = NULL, var_type = c("mad", "IQR",
  "sd"), prop = 0.95, set_shape = c("left", "centre"), n_perc = 1000)
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', most likely corresponding to lethality scores
produced by \code{\link{get.leth.scores}}. It is also possible to normalize re-scaled screen
viability data, such as yielded by \code{\link{read.screen.data}}, or by \code{\link{combine.screens}},
although these tend to be less comparable between cell lines and replicates than the scores.}

\item{my_gamma}{scalar or numeric vector, with value(s) between 0 and a finite
number. Its default value is \code{NULL}. If given, it will be used to construct the
core sets by using the relative distance between negative and positive
controls' distributions. If a numeric vector is given, it must contain as many
entries as columns in the data_only slot of data_rscreen. A value of 1 means that
the core set includes all lethality scores up to the ratio between the MAD
of negative controls, and the sum of the MADs of negative and positive controls,
assuming the robust variability measure MAD is used.}

\item{var_type}{string indicating the statistic used as variability measure.
Possible values are 'mad' when the median absolute deviation (MAD) is used
(the default), 'IQR' when the inter-quartile range is used, and "sd" when the
standard deviation is used.}

\item{prop}{scalar between 0 and 1, corresponding to the desired proportion of
lethality scores to be included in the core set, per replicate. The default
value is 0.95. If \code{my_gamma} is numeric, any value for this argument is ignored.}

\item{set_shape}{string indicating the shape of the core set. It accepts
"left" (the default), when all scores to the left of the set threshold are
included, per replicate, and "centre", when scores included in the core
set only exclude the largest and smallest ones, at equal frequencies.}

\item{n_perc}{number of percentiles to be used to yield a representation of each
core set. The default value is 1000, which is appropriate for pooled, whole
genome screens. For small, hit-picking screens a smaller number may be more appropriate.
The user may check the core set sizes to decide if in doubt. Note that the value
of \code{n_perc}  should not be larger than that of any core set.}
}
\value{
An object of class rscreen.object, with its 'data_only' slot containing quantile-normalized
lethality scores.
}
\description{
This is a wrapper for multiple functions, that will determine core sets and
quantile-normalize scores for all replicates in a dataset.
}
\examples{
# See vignette
}
\seealso{
\code{\link{read.screen.data}} on reading screen data, \code{\link{combine.screens}}
to combine data from multiple screens into a single object, \code{\link{get.inv.set}} for
obtaining core sets for all replicates in the dataset and \code{\link{get.leth.scores}}
to compute lethality scores.
}
\concept{
scores screen siRNA CRISPR CRISPR-Cas9 crispr-cas9 CRISPR/Cas9 crispr/cas9 quantile normalization
}
