% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_set.R
\name{get.th.inv.set}
\alias{get.th.inv.set}
\alias{getThInvSet}
\title{Compute thresholds that define core sets per replicate}
\usage{
get.th.inv.set(data_rscreen, my_gamma = NULL, var_type = c("mad", "IQR",
  "sd"), prop = 0.95, set_shape = c("left", "centre"))
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', either yielded by \code{\link{read.screen.data}}, or
by \code{\link{combine.screens}}. Typically the 'data_only' slot of this object
contains lethality scores computed by \code{\link{get.leth.scores}}, which already
reflect a value relative to negative and positive controls, per replicate.}

\item{my_gamma}{scalar or numeric vector, with value(s) between 0 and a finite
number. Its default value is \code{NULL}. If given, it will be used to construct the
core sets by using the relative distance between negative and positive
controls' distributions. If a numeric vector is given, it must contain as many
entries as columns in the data_only slot of data_rscreen. A value of 1 means that
the core set includes all lethality scores up to the ratio between the MAD
of negative controls, and the sum of the MADs of negative and positive controls,
assuming the robust variability measure MAD is used.}

\item{var_type}{string indicating the statistic used as variability measure.
Possible values are 'mad' when the median absolute deviation (MAD) is used
(the default), 'IQR' when the inter-quartile range is used, and "sd" when the
standard deviation is used.}

\item{prop}{scalar between 0 and 1, corresponding to the desired proportion of
lethality scores to be included in the core set, per replicate. The default
value is 0.95. If \code{my_gamma} is numeric, any value for this argument is ignored.}

\item{set_shape}{string indicating the shape of the core set. It accepts
"left" (the default), when all scores to the left of the set threshold are
included, per replicate, and "centre", when scores included in the core
set only exclude the largest and smallest ones, at equal frequencies.}
}
\value{
A list with the following elements: the first with the thresholds defining
the core sets (one value or pair per sample, so either a vector or a matrix),
}
\description{
This is an internal function to compute thresholds
to be used later on for constructing
core sets. Two methods are available: using the distance between negative
and positive controls to determine the core set, and including in the set
a proportion of the lethality scores. If no well annotation is available indicating
which wells correspond to assay controls, then core
sets are computed using a proportion of all available features.
To be used internally, it is called by
get.inv.set.
}
\seealso{
\code{\link{get.inv.set}} the wrapper that calls this functions and others
to yield the resulting core sets, \code{\link{get.rscreenorm}} to normalize data
from multiple screens using core sets and \code{\link{get.leth.scores}} to compute
scores that make observations in different screens comparable.
}
\concept{
core set normalization quantile threshold
}
