% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_set.R
\name{get.wtype}
\alias{get.wtype}
\alias{getWtype}
\title{Extract well type from data_ann and format it as matrix if need be}
\usage{
get.wtype(data_only, data_ann)
}
\arguments{
\item{data_only}{data matrix containing lethality scores, per replicate.}

\item{data_ann}{matrix or list}
}
\value{
A list with the following elements: the first with the thresholds defining
the core sets (one value or pair per sample, so either a vector or a matrix),
}
\description{
This is an internal function that
extracts well type info from 'data_ann' slot. This function is
called by \code{\link{get.th.inv.set}}, \code{\link{sel.scores.inv.set}} and others.
}
\seealso{
\code{\link{get.inv.set}} the wrapper that calls this functions and others
to yield the resulting core sets, \code{\link{get.rscreenorm}} to normalize data
from multiple screens using core sets and \code{\link{get.leth.scores}} to compute
scores that make observations in different screens comparable.
}
\concept{
core set normalization quantile threshold
}
