% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{hasWtype}
\alias{hasWtype}
\title{Checks if an rscreen.object has a well type variable}
\usage{
hasWtype(data_rscreen)
}
\arguments{
\item{data_rscreen}{an rscreen.object}
}
\value{
A logical value that is TRUE if a well type variable is found.
}
\description{
To be used internally, this function checks for a well type variable
within the rscreen.object
}
