% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qnorm.R
\name{mypred.lm}
\alias{mypred.lm}
\title{Given a fit object and a new data, yields predictions of the fit object for all obs in new data
by using the predict() within the data range, and a shift outside of this range (so this
involves a piecewise-linear prediction rule)}
\usage{
mypred.lm(my_fit, newdata, mymin_pm, mymax_pm)
}
\arguments{
\item{my_fit}{an object of class "lm", generated by a call to \code{\link{lm}}. It is expected
to represent the regression of the observed percentiles to their quantile-normalized values.}

\item{newdata}{numeric vector containing all lethality scores, for which predictions will be made
based upon the fit given by \code{my_fit}}

\item{mymin_pm}{numeric (scalar), the minimum value for the core set lethality scores}

\item{mymax_pm}{numeric (scalar), the maximum value for the core set lethality scores}
}
\value{
a numeric vector with as many entries as \code{newdata},containing the predicted values,
which correspond to the quantile-normalized lethality scores for all observed scores.
}
\description{
This is an internal function that computes fitted values for observed values in the core set range,
and for those outside it uses a extrapolation equivalent to a shift. It is called by
\code{\link{get.all.qnorm}} and is to be used internally, so has no functionality other
than what is described.
}
\seealso{
\code{\link{get.rscreenorm}} which yields core sets and quantile-normalizes data at
once, code{\link{get.qnorm}} which quantile-normalizes data for a given core set and
\code{\link{get.all.qnorm}} to compute the quantile normalized scores.
}
