% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{mysplit}
\alias{mysplit}
\title{Splits and extracts single value from each entry in a vector of strings}
\usage{
mysplit(xi, char.vector, split.by, result.sel, fixed = FALSE)
}
\arguments{
\item{xi}{the element to be selected, typically is a number varying from 1 to the length of the character vector.}

\item{char.vector}{the character vector containing the strings to be split.}

\item{split.by}{the string to be used to split each entry in char.vector - see \code{\link{strsplit}} for details.}

\item{result.sel}{integer indicating the index of the element after splitting to be selected.}

\item{fixed}{logical, to be passed on to \code{\link{strsplit}}.
\code{\link{strsplit}} returns  a list of results, of which the one with index result.sel is selected.}
}
\value{
A  single string.
}
\description{
To be used internally, this function splits a string vector using a given separator,
extracts one of the results and combines all extracted results in a vector. The split
is done by using \code{\link{strsplit}}. This function
was written to be used sequentially on a vector via \code{\link{sapply}}.
}
\examples{
xs <- paste(LETTERS[1:5], 5:1, sep="_")
xs
sapply(1:length(xs), mysplit, char.vector = xs, split.by = "_", result.sel = 1)
# Compare it with the result from directly applying strsplit:
strsplit(xs, "_")
}
