% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{nr.screen}
\alias{nr.screen}
\title{Gets the number of features in the rscreen.object}
\usage{
nr.screen(data_rscreen)
}
\arguments{
\item{data_rscreen}{an rscreen.object}
}
\value{
A numeric value.
}
\description{
To be used internally, this function extracts the number of rows
in the data_only slot of  the rscreen.object
}
