% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{nwt}
\alias{nwt}
\title{Extracts the number of well types}
\usage{
nwt(data_rscreen)
}
\arguments{
\item{data_rscreen}{an rscreen.object}
}
\value{
A numeric value.
}
\description{
To be used internally, this function extracts the number of levels
in the well type variable within a given rscreen.object. If the object
has no well type, it will return 0.
}
