% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_screen_data.R
\name{read.screen.data}
\alias{read.screen.data}
\alias{readScreenData}
\title{Reads screen data from a tab-delimited file}
\usage{
read.screen.data(filename, mydir = NULL, sep = "\\t", check_names = FALSE,
  ann_cols_names = NULL, n_ann_cols = length(ann_cols_names),
  wtype_col = NULL, wtype_lab = c("sample", "pos", "neg"),
  use_plate = FALSE, plate_var = "plate")
}
\arguments{
\item{filename}{string, the name of the tab-delimited file containing the data
to be read, including the file extension.}

\item{mydir}{string, indicating the complete folder path where the data file
is to be found. It is assumed it ends without any (forward or backward) slashes, for example
'/home/projects' instead of '/home/projects/'. If \code{NULL}, it will get the current
work directory via \code{\link{getwd}}.}

\item{sep}{the separator to be used while reading in the data file,
containing a value accepted by read.delim.}

\item{check_names}{logical, indicating whether or not the column names in the data
 file need to be checked for consistency or not. To be used while reading
 in the data file. Defaults to \code{FALSE}, when comparability between
column names in the data and in the phenotypic data table is guaranteed.}

\item{ann_cols_names}{vector of strings containing the names of the annotation
columns. It is assumed that these columns are the first few ones in the data file.}

\item{n_ann_cols}{an integer giving the number of annotation columns (from the
first one) in the data read. Any value given is only used if ann_cols_names is
\code{NULL}, otherwise it is ignored. If ann_cols_names is given and n_ann_cols
is left empty, n_ann_cols is by default the length of ann_cols_names.}

\item{wtype_col}{string with the name of the annnotation column corresponding
to the well annotation column. If not present, the function issues a warning.
For lethality scores to be computed, this column must exist.}

\item{wtype_lab}{labels in wtype_col variable corresponding to library/sample
features, positive and negative controls,
respectively. Defaults to 'c("sample","pos", "neg")'.}

\item{use_plate}{logical indicating whether or not the data contains a plate variable.
Defaults to \code{FALSE}. If \code{TRUE}, the plate will be used when computing the lethality scores
later on with \code{\link{get.leth.scores}}.}

\item{plate_var}{string with the name of the plate variable.
Only needed if use_plate is \code{TRUE}. Defaults to "plate".}
}
\value{
An object of class rscreen.object.
}
\description{
\code{read.screen.data} reads in a tab-delimited file containing annotation
in the first k columns and screen data per replicate in the remaining columns.
Annotation columns include gene (target) id as well as feature/well type, where library
features, positive and negative controls are identified. Arrayed screens
typically also include a plate identifier. All data columns are assumed to have
the same annotation. If that is not the case, the data is assumed to correspond to
different screens, and it must be loaded for each separate screen first before combining
it into a single object using \code{\link{combine.screens}}.
}
\examples{
# See vignette
}
\seealso{
\code{\link{combine.screens}} to combine data from multiple screens,
\code{\link{combine.pdata}} to combine pdata from multiple screens into a single object,
and \code{\link{get.leth.scores}} to compute scores that make observations in
different replicates and screens comparable.
}
\concept{
input screen siRNA CRISPR CRISPR-Cas9 crispr-cas9 CRISPR/Cas9 crispr/cas9
}
