% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{scr.ann}
\alias{scr.ann}
\title{Extracts the annotation data of an rscreen object}
\usage{
scr.ann(data_rscreen)
}
\arguments{
\item{data_rscreen}{an rscreen.object}
}
\value{
A data frame containing annotation columns.
}
\description{
This function extracts the  'data_only' slot of  a given rscreen.object.
}
