% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{scr.data}
\alias{scr.data}
\title{Extracts the screen data of an rscreen object}
\usage{
scr.data(data_rscreen)
}
\arguments{
\item{data_rscreen}{an rscreen.object}
}
\value{
A numeric data matrix.
}
\description{
This function extracts the  'data_only' slot of  a given rscreen.object.
}
