% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{scr.names}
\alias{scr.names}
\title{Extracts sample names of an rscreen object}
\usage{
scr.names(data_rscreen)
}
\arguments{
\item{data_rscreen}{an rscreen.object}
}
\value{
A character vector.
}
\description{
This function extracts the column names
of the 'data_only' slot of  a given rscreen.object. If 'the argument 'data_rscreen'
is not of class 'rscreen.object', the column names of the object are returned.
}
