% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{sel.rep}
\alias{sel.rep}
\title{Selects densities to be used.}
\usage{
sel.rep(mydens, xi)
}
\arguments{
\item{mydens}{a matrix of which columns are to be selected}

\item{xi}{the index of the column to be selected}
}
\value{
The density selected
}
\description{
To be used internally, this function selects columns of a density matrix.
}
\seealso{
\code{\link{pdmw}}
}
