% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_set.R
\name{sel.scores.inv.set}
\alias{sel.scores.inv.set}
\alias{selScoresInvSet}
\title{Creates an object indicating observations in core set}
\usage{
sel.scores.inv.set(data_rscreen, th_obj)
}
\arguments{
\item{data_rscreen}{an object of class 'rscreen.object', either yielded by \code{\link{read.screen.data}}, or
by \code{\link{combine.screens}}. Typically the 'data_only' slot of this object
contains lethality scores computed by \code{\link{get.leth.scores}}, which already
reflect a value relative to negative and positive controls, per replicate.}

\item{th_obj}{object containing thresholds per replicate, as returned by
get.th.inv.set}
}
\value{
A logical matrix with the same number of columns as the data_only slot of data_rscreen,
and as many rows as annotated with "sample" in the data_only slot
}
\description{
This is an internal function that
uses data_rscreen and a threshold object and returns a
logical matrix, with the same dimensions as data_only, with \code{TRUE} for every
lethality score in the core set, and \code{FALSE} for all others. To be used
internally, it is called by get.inv.set.
}
\seealso{
\code{\link{get.inv.set}} the wrapper that calls this functions and others
to yield the resulting core sets, \code{\link{get.th.inv.set}} that produces the
thresholds per replicate, \code{\link{get.rscreenorm}} to normalize data from multiple
screens using core sets and \code{\link{get.leth.scores}} to compute scores that
make observations in different screens comparable.
}
\concept{
core set normalization quantile threshold
}
